@@ -6,9 +6,10 @@
     "@ant-design/icons": "^4.1.0",
     "@craco/craco": "^7.1.0",
     "@fortawesome/free-brands-svg-icons": "^5.14.0",
-    "@testing-library/jest-dom": "^4.2.4",
-    "@testing-library/react": "^9.3.2",
-    "@testing-library/user-event": "^7.1.2",
+    "@testing-library/dom": "^10.4.1",
+    "@testing-library/jest-dom": "^6.9.1",
+    "@testing-library/react": "^16.3.0",
+    "@testing-library/user-event": "^14.6.1",
     "amazon-cognito-identity-js": "^4.2.3",
     "antd": "^4.2.2",
     "aws-amplify": "^4.3.12",
@@ -16,15 +17,15 @@
     "dayjs": "^1.8.27",
     "history": "^4.10.1",
     "ky": "^0.20.0",
-    "react": "^16.13.1",
-    "react-dom": "^16.13.1",
+    "react": "18.3.1",
+    "react-dom": "18.3.1",
     "react-ga": "^3.1.2",
     "react-icons": "^3.11.0",
-    "react-redux": "^7.2.0",
+    "react-redux": "^9.2.0",
     "react-router-dom": "^5.2.0",
-    "react-scripts": "5.0.0",
+    "react-scripts": "^5.0.1",
     "react-star-rating-component": "^1.4.1",
-    "redux": "^4.0.5",
+    "redux": "^5.0.1",
     "redux-devtools-extension": "^2.13.8",
     "redux-logger": "^3.0.6",
     "redux-thunk": "^2.3.0",
@@ -54,6 +55,7 @@
   },
   "proxy": "http://localhost:5001",
   "devDependencies": {
+    "@babel/plugin-proposal-private-property-in-object": "^7.21.11",
     "npm-run-all": "^4.1.5",
     "rimraf": "^3.0.2"
   }
