@@ -1,5 +1,6 @@
 import React from 'react';
-import ReactDOM from 'react-dom';
+import { createRoot } from 'react-dom/client';
+
 import { BrowserRouter as Router, Route } from 'react-router-dom'
 import './index.css';
 import App from './App';
@@ -14,14 +15,16 @@ import Amplify from 'aws-amplify'
 import awsmobile from './awsmobile'
 Amplify.configure(awsmobile)
 
-ReactDOM.render(
-  
+
+const container = document.getElementById('root');
+const root = createRoot(container);
+
+root.render(
   <Provider store={store}>
     <Router>
       <Route component={App} />
     </Router>
-  </Provider>,
-  document.getElementById('root')
+  </Provider>
 );
 
 // If you want your app to work offline and load faster, you can change
