@@ -2,24 +2,8 @@
 // allows you to do things like:
 // expect(element).toHaveTextContent(/react/i)
 // learn more: https://github.com/testing-library/jest-dom
-import '@testing-library/jest-dom/extend-expect';
-// Object.defineProperty(window, 'matchMedia', {
-//     writable: true,
-//     value: jest.fn().mockImplementation(query => ({
-//       matches: false,
-//       media: query,
-//       onchange: null,
-//       addListener: jest.fn(), // deprecated
-//       removeListener: jest.fn(), // deprecated
-//       addEventListener: jest.fn(),
-//       removeEventListener: jest.fn(),
-//       dispatchEvent: jest.fn(),
-//     })),
-//   });
+import '@testing-library/jest-dom';
+console.log('setupTests.js is running');
 
-global.matchMedia = global.matchMedia || function () {
-  return {
-    addListener: jest.fn(),
-    removeListener: jest.fn(),
-  };
-};
+
+
